refactor(sync): extract joiner pairing client into peers/join.rs
